rename Member to Set, only provide project func
The Member typeclass provided value-level proof for the existence of a
type a Set. However, it was implemented to not type check for types that
aren't present. I see this as the useful behaviour, so I renamed it and
only export the `project` function that returns the value at that type.

The alternate solution is to add a False case in the instance
definition, which would move the "error" from type to value level.